Make shard allocation give preference to current location
In this implementation, it gives infinite preference to the current location.
This works well in most if not all recent migrations that I am aware of,
but will work poorly in the special case of simply adding a new node,
without removing any nodes or increasing the number of copies. In this case
all shards will stay where they are and no shards will end up on the new nodes.

I think this is still a huge practical improvement over the former algorithm.
In the future if this becomes a concern, I think a potential solution is basically
to do this first, and then go through a secondary rebalancing step,
where the preference for current location is loosened and shards are moved from
fuller nodes to emptier nodes until an acceptable uniformity is reached.